Security update __init__.py
Secret key was hardcoded to "dev".  A proper secret key shall be added to .env to make it safe and publicly deployable. If it is not present it'll fallback to "dev" so it will work even if you don't do it

sqlite path also changed to make it more secure, database URL can be also added secretly.